Windows: use MAXDWORD - 1 for ReadTotalTimeoutConstant for the Blocking case.
The MSDN documentation is unclear on what happens when it's set to
MAXDWORD.  I don't have the capability to test across different
windows versions with multiple serial devices, so this seems to be the
safer option since this is what other libraries seem to do.

For example here:
http://forums.ni.com/attachments/ni/140/18925/1/SerialENETWrite.cpp

This link suggests another option for Blocking:
http://stackoverflow.com/questions/10390151/how-do-i-use-commtimeouts-to-wait-until-bytes-are-available-but-read-more-than-o

The documentation is poor about these different corner cases, but
MAXDWORD-1 seems to be a fairly safe choice.